add guards for file backup

+++ index.js
@@ -34,10 +34,10 @@ function initialSetup(firstTime) {
     if (!firstTime) {
         // Migration from pre-1.6
         // Back everything up, just in case
-        fs.copySync(configPath, configPath + ".bak");
-        fs.copySync(serversPath, serversPath + ".bak");
-        fs.copySync(versionsPath, versionsPath + ".bak");
-        fs.copySync(hashPath, hashPath + ".bak");
+        if (fs.existsSync(configPath)) fs.copySync(configPath, configPath + ".bak");
+        if (fs.existsSync(serversPath)) fs.copySync(serversPath, serversPath + ".bak");
+        if (fs.existsSync(versionsPath)) fs.copySync(versionsPath, versionsPath + ".bak");
+        if (fs.existsSync(hashPath)) fs.copySync(hashPath, hashPath + ".bak");
     } else {
         // First-time setup
         // Copy default servers
